SoundSystem: fix delay
There was some delay between video and sound. It did not exist with oss
system so it is not because of sound speed (< light speed).

Some changes in sound system:

Use hw:0,0 device and so we directly open the sound card.

Use snd_pcm_hw_params_set_periods and snd_pcm_hw_params_set_buffer_size_near: I
don't know what it is for but it sound better.

Exception improvement: not essential but better.